pyproject.toml
[tool.poetry]
name = "cinema"
version = "0.1.0"
description = ""
authors = ["yaaspichikhin <yaaspichikhin@rt-dc.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
faker = "^22.0.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
